Avoid implicit type conversions

+++ ThrEd4/ThrEdTypes.h
@@ -1246,10 +1246,10 @@ inline fPOINTATTR::fPOINTATTR(double rhsX, double rhsY, uint32_t rhsA) noexcept 
   attribute = rhsA;
 }
 
-inline dPOINT::dPOINT(fPOINT const& rhs) noexcept : x(rhs.x), y(rhs.y) {
+inline dPOINT::dPOINT(fPOINT const& rhs) noexcept : x(gsl::narrow_cast<double>(rhs.x)), y(gsl::narrow_cast<double>(rhs.y)) {
 }
 
-inline dPOINT::dPOINT(float rhsX, float rhsY) noexcept : x(rhsX), y(rhsY) {
+inline dPOINT::dPOINT(float rhsX, float rhsY) noexcept : x(gsl::narrow_cast<double>(rhsX)), y(gsl::narrow_cast<double>(rhsY)) {
 }
 
 inline dPOINT::dPOINT(double rhsX, double rhsY) noexcept : x(rhsX), y(rhsY) {
@@ -1259,14 +1259,14 @@ inline auto dPOINT::operator==(dPOINT const& rhs) const noexcept -> bool {
 }
 
 inline auto dPOINT::operator=(fPOINT const& rhs) noexcept -> dPOINT& {
-  x = rhs.x;
-  y = rhs.y;
+  x = gsl::narrow_cast<double>(rhs.x);
+  y = gsl::narrow_cast<double>(rhs.y);
   return *this;
 }
 
 inline auto dPOINT::operator=(BSEQPNT const& rhs) noexcept -> dPOINT& {
-  x = rhs.x;
-  y = rhs.y;
+  x = gsl::narrow_cast<double>(rhs.x);
+  y = gsl::narrow_cast<double>(rhs.y);
   return *this;
 }
